+++ crankshaft-docker/src/lib.rs
@@ -12,7 +12,9 @@ pub mod service;
 
 use bollard::secret::Node;
 use bollard::secret::SystemInfo;
+use crankshaft_config::backend::docker::EventConfig;
 use crankshaft_events::Event;
+use crankshaft_events::TaskId;
 use thiserror::Error;
 use tokio::sync::broadcast;
 
@@ -93,8 +95,9 @@ impl Docker {
         &self,
         image: impl Into<String>,
         token: tokio_util::sync::CancellationToken,
+        events_ctx: Option<(broadcast::Sender<Event>, TaskId)>,
     ) -> Result<Option<()>> {
-        ensure_image(self, image, token).await
+        ensure_image(self, image, token, events_ctx).await
     }
 
     /// Removes an image from the Docker daemon.
@@ -178,9 +181,11 @@ pub struct EventOptions {
     /// The sender for sending events.
     pub sender: broadcast::Sender<Event>,
     /// The task id for the events.
-    pub task_id: u64,
+    pub task_id: TaskId,
     /// Whether or not send the task started event.
     pub send_start: bool,
+    /// User-controlled event configuration.
+    pub user_config: EventConfig,
 }
 
 #[cfg(test)]