+++ crankshaft-monitor/src/service.rs
@@ -6,6 +6,7 @@ use std::sync::Arc;
 use std::time::SystemTime;
 
 use crankshaft_events::Event as CrankshaftEvent;
+use crankshaft_events::TaskId;
 use futures_core::Stream;
 use tokio::sync::RwLock;
 use tokio::sync::broadcast;
@@ -22,6 +23,9 @@ use crate::proto::CancelTaskRequest;
 use crate::proto::CancelTaskResponse;
 use crate::proto::Event;
 use crate::proto::ExitStatus;
+use crate::proto::ImagePullFailedEvent;
+use crate::proto::ImagePullFinishedEvent;
+use crate::proto::ImagePullStartedEvent;
 use crate::proto::ServiceStateRequest;
 use crate::proto::ServiceStateResponse;
 use crate::proto::SubscribeEventsRequest;
@@ -39,9 +43,6 @@ use crate::proto::TaskStdoutEvent;
 use crate::proto::event::EventKind;
 use crate::proto::exit_status::ExitStatusKind;
 use crate::proto::monitor_server::Monitor;
-
-/// Represents a task identifier.
-pub type TaskId = u64;
 
 /// Helper trait for converting Crankshaft types into Protobuf types.
 trait IntoProtobuf<T> {
@@ -118,6 +119,15 @@ impl IntoProtobuf<EventKind> for CrankshaftEvent {
                 id,
                 message: message.to_vec(),
             }),
+            CrankshaftEvent::ImagePullStarted { id, name } => {
+                EventKind::ImagePullStarted(ImagePullStartedEvent { id, name })
+            }
+            CrankshaftEvent::ImagePullFailed { id, name, message } => {
+                EventKind::ImagePullFailed(ImagePullFailedEvent { id, name, message })
+            }
+            CrankshaftEvent::ImagePullFinished { id, name } => {
+                EventKind::ImagePullFinished(ImagePullFinishedEvent { id, name })
+            }
         }
     }
 }
@@ -177,6 +187,7 @@ impl MonitorService {
                 r = events.recv() => match r {
                     Ok(event) => {
                         let (id, remove) = match event {
+                            CrankshaftEvent::ImagePullStarted { .. } | CrankshaftEvent::ImagePullFailed { .. } | CrankshaftEvent::ImagePullFinished { .. } => continue,
                             CrankshaftEvent::TaskCreated { id, .. } |
                             CrankshaftEvent::TaskStarted { id }
                             | CrankshaftEvent::TaskContainerCreated { id, .. }